feat: Tag cached_issues in db with a version of the pydantic model
